Pr moving average (#10)

+++ Src/periphery/as5600/as5600.hpp
@@ -41,8 +41,7 @@
 typedef uint8_t as5600_addr;
 
 
-typedef enum as5600_error_e
-{
+typedef enum as5600_error_e {
         AS5600_SUCCESS = 0,
         AS5600_BAD_PARAMETER,
         AS5600_RUNTIME_ERROR,
@@ -50,8 +49,7 @@ typedef enum as5600_error_e
 } as5600_error_t;
 
 
-struct as5600_data
-{
+struct as5600_data {
     bool dir; // 0 clockwise | 1 counterclockwise 
     uint16_t raw_angle;
     uint16_t max_angle;
@@ -59,18 +57,16 @@ struct as5600_data
     uint16_t angle;
     uint8_t mag_status;
     uint8_t min_angle_deg = 18;
-    // uint32_t min_angle_steps = 206;
 };
 
 void wait(uint8_t time_ns);
 
-class As5600Periphery
-{
+class As5600Periphery {
 public:
     // TODO: add sensor parameters
     as5600_data data = {.dir=0,.raw_angle = 0, .max_angle = 4, .start_angle=0, .angle=400, .mag_status=0};
     
-    as5600_error_t init(uint32_t min_angle, uint32_t max_angle);
+    as5600_error_t init(uint16_t min_angle, uint16_t max_angle);
     int8_t reset();
     int8_t update_zpos();
     as5600_error_t calibrate();
